+++ lambda-example/src/main/scala/zio/lambda/example/CustomEvent.scala
