+++ lambda-example/src/main/scala/zio/lambda/example/CustomResponse.scala
